test
test

+++ Classes/Marketplace.swift
@@ -58,7 +58,7 @@ public class MarketplaceViewController: UIViewController, WKUIDelegate {
         let topPadding = self.view.safeAreaLayoutGuide
         let topBarLeft = CGRect(x: 0.0,
                             //y: self.view.frame.maxY - (self.view.frame.height),
-                                y: topPadding,
+                                y: CGFloat(topPadding),
                             width: self.view.frame.size.width/2,
                             height: self.view.frame.height/CGFloat(sizeMetric))
